pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "LangGraph agent project with a uv-managed Jupyter workflow."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0,<2.0",
    "langchain-core>=1.0,<2.0",
    "langchain-openai>=1.1.15",
    "langgraph>=1.0,<2.0",
    "langsmith>=0.3.0,<1.0",
    "pygraphviz>=1.14",
]

[dependency-groups]
dev = [
    "harbor>=0.4.0",
    "ipykernel>=6.29.5,<7",
    "jupyterlab>=4.4,<5",
    "python-dotenv>=1.0,<2",
]
